+++ main/utils/payload.h
@@ -2,25 +2,26 @@
 #define PAYLOAD_H
 
 #include <string>
+#include <cstdint>
 
 class Payload {
 private:
     char *buffer;
-    size_t len;
+    uint16_t len;
 
     Payload &operator=(const Payload &other) = delete;
     Payload(const Payload &other) = delete;
 
 public:
     Payload(const std::string &as_string);
-    Payload(size_t length = 0);
-    Payload(char *buffer, size_t length);
+    Payload(uint16_t length = 0);
+    Payload(char *buffer, uint16_t length);
     Payload(Payload &&other);
     ~Payload();
     
     Payload &operator=(Payload &&other);
     Payload &operator+=(const Payload &other);
-    size_t length() const;
+    uint16_t length() const;
     const char *data() const;
 };
